Renamed few files and internal classes for more clarity
